Moved slope/aspect calc methods to gradients.py.
